+++ services/api/src/roles/roles.module.ts
@@ -4,6 +4,7 @@ import { RolesController } from './roles.controller';
 import { RolesService } from './roles.service';
 import { HostProfile } from './entities/host-profile.entity';
 import { CleanerProfile } from './entities/cleaner-profile.entity';
+import { User } from '../auth/entities/user.entity';
 import { AuthModule } from '../auth/auth.module';
 
 /**
@@ -14,7 +15,7 @@ import { AuthModule } from '../auth/auth.module';
  */
 @Module({
   imports: [
-    TypeOrmModule.forFeature([HostProfile, CleanerProfile]),
+    TypeOrmModule.forFeature([User, HostProfile, CleanerProfile]),
     AuthModule,
   ],
   controllers: [RolesController],